+++ pkg/notify/notify.go
@@ -14,6 +14,7 @@ package notify
 
 import (
 	"bytes"
+	"crypto/sha512"
 	"fmt"
 	"io/ioutil"
 	"reflect"
@@ -58,13 +59,13 @@ func NewReceiver(logger log.Logger, c *config.ReceiverConfig, t *template.Templa
 }
 
 // Notify manages JIRA issues based on alertmanager webhook notify message.
-func (r *Receiver) Notify(data *alertmanager.Data) (bool, error) {
+func (r *Receiver) Notify(data *alertmanager.Data, hashJiraLabel bool) (bool, error) {
 	project, err := r.tmpl.Execute(r.conf.Project, data)
 	if err != nil {
 		return false, errors.Wrap(err, "generate project from template")
 	}
 
-	issueGroupLabel := toGroupTicketLabel(data.GroupLabels)
+	issueGroupLabel := toGroupTicketLabel(data.GroupLabels, hashJiraLabel)
 
 	issue, retry, err := r.findIssueToReuse(project, issueGroupLabel)
 	if err != nil {
@@ -232,8 +233,23 @@ func deepCopyWithTemplate(value interface{}, tmpl *template.Template, data inter
 
 // toGroupTicketLabel returns the group labels as a single string.
 // This is used to reference each ticket groups.
-// String is the form of an ALERT Prometheus metric name, with all spaces removed.
-func toGroupTicketLabel(groupLabels alertmanager.KV) string {
+// (old) default behavior: String is the form of an ALERT Prometheus metric name, with all spaces removed.
+// new opt-in behavior: String is the form of JIRALERT{sha512hash(groupLabels)}
+// hashing ensures that JIRA validation still accepts the output even
+// if the combined length of all groupLabel key-value pairs would be
+// longer than 255 chars
+func toGroupTicketLabel(groupLabels alertmanager.KV, hashJiraLabel bool) string {
+	// new opt in behavior
+	if hashJiraLabel {
+		hash := sha512.New()
+		for _, p := range groupLabels.SortedPairs() {
+			kvString := fmt.Sprintf("%s=%q,", p.Name, p.Value)
+			_, _ = hash.Write([]byte(kvString)) // hash.Write can never return an error
+		}
+		return fmt.Sprintf("JIRALERT{%x}", hash.Sum(nil))
+	}
+
+	// old default behavior
 	buf := bytes.NewBufferString("ALERT{")
 	for _, p := range groupLabels.SortedPairs() {
 		buf.WriteString(p.Name)